+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 * Update :command:`fuc-find` command to allow users to control whether to use recursive retrieving.
 * Add new command :command:`ngs-trim`.
 * Add new command :command:`ngs-quant`.
+* Add new submodule ``pykallisto``.
 
 0.29.0 (2021-12-19)
 -------------------